separate clientside socket logic
