+++ client/index.js
@@ -1,52 +1,10 @@
-import io from 'socket.io-client';
-
-const socket = io();
+import imageGameLogic from './sockets/imageGame';
+import waitingRoomLogic from './sockets/waitingRoom';
 
 if (window.location.pathname === '/waitingroom') {
-  const form = document.getElementById('form');
-
-  form.onsubmit = (e) => {
-    const m = document.getElementById('m');
-    e.preventDefault(); // prevents page reloading
-    console.log('message', m.value);
-    socket.emit('chat message', m.value);
-    m.value = '';
-    return false;
-  };
-
-  socket.on('chat message', (msg) => {
-    const message = document.createElement('li');
-    message.innerText = msg;
-    document.getElementById('messages').append(message);
-  });
+  waitingRoomLogic();
 }
 
 if (window.location.pathname === '/imagegame') {
-  const clickableImages = Array.from(document.getElementsByClassName('clickable'));
-  const receivedImage = document.getElementById('received');
-  const matchResult = document.getElementById('matchResult');
-
-  clickableImages.forEach((image) => {
-    image.addEventListener('click', (event) => {
-      socket.emit('imageClicked', {
-        data: getBase64Image(event.target),
-      });
-    });
-  });
-
-  socket.on('imageClicked', (data) => {
-    console.log(data);
-    receivedImage.src = data.data;
-    matchResult.innerText = `Match percentage: ${data.percent}%`;
-  });
-}
-
-function getBase64Image(img) {
-  const c = document.createElement('canvas');
-  c.height = img.naturalHeight;
-  c.width = img.naturalWidth;
-  const ctx = c.getContext('2d');
-
-  ctx.drawImage(img, 0, 0, c.width, c.height);
-  return c.toDataURL();
+  imageGameLogic();
 }
