--- server/rooms/waitingRoom.js
@@ -1,2 +1,0 @@
-
-
